Move healthcheck tests to separate file
